izmjene1

+++ backend/controller/user/deleteAddToCartProduct.js
@@ -2,10 +2,10 @@ const addToCartModel = require("../../models/cartProduct")
 
 const deleteAddToCartProduct = async(req,res)=>{
     try {
-        const currentUserId = req.currentUserId
+        const currentUserId = req.userId
         const addToCartProductId = req.body._id
 
-        const deleteProduct = await addToCartModel.deleteOne({_id : addToCartProductId})
+        const deleteProduct = await addToCartModel.deleteOne({_id : addToCartProductId, userId : currentUserId})
 
         res.json({
             messege : "Product deleted from cart",
